scenario/flask/requirements.txt
kafka-python
numpy
pyelasticsearch
pymongo
requests
sklearn
iso8601
cassandra-driver
